Tehtävät 4.18, 4.19, 4.20, 4.21

+++ controllers/blogs.js
@@ -3,14 +3,6 @@ const blogRouter = require('express').Router()
 const Blog = require('../models/blog')
 const User = require('../models/user')
 
-const getTokenFrom = request => {
-  const authorization = request.get('authorization')
-  if (authorization && authorization.toLowerCase().startsWith('bearer ')) {
-    return authorization.substring(7)
-  }
-  return null
-}
-
 blogRouter.get('/', async (request, response) => {
   const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })  
   response.json(blogs)
@@ -29,13 +21,13 @@ blogRouter.get('/:id', async (request, response, next) => {
 })
 
 blogRouter.post('/', async (request, response, next) => {
-  
     const body = request.body
+    //const token = getTokenFrom(request)
+    //const decodedToken = jwt.verify(token, process.env.SECRET)
 
-    const token = getTokenFrom(request)
-    const decodedToken = jwt.verify(token, process.env.SECRET)
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
 
-    if (!token || !decodedToken.id) {
+    if (!request.token || !decodedToken.id) {
       return response.status(401).json({ error: 'token missing or invalid' })
     }
     const user = await User.findById(decodedToken.id)
@@ -61,24 +53,47 @@ blogRouter.post('/', async (request, response, next) => {
     response.json(savedBlog.toJSON())
 })
 
-blogRouter.put('/:id', async (request, response, next) => {
-    const body = request.body
+blogRouter.delete('/:id', async (request, response, next) => {
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
 
-    const blog = {
-      title: body.title,
-      author: body.author,
-      url: body.url,
-      likes: body.likes,
-      id: body.id
+    if (!request.token || !decodedToken.id) {
+      return response.status(401).json({ error: 'token missing or invalid' })
     }
 
-    await Blog.findByIdAndUpdate(request.params.id, blog)
-    response.status(200).end()
-})
+    const blogToDelete = await Blog.findById(request.params.id)
 
-blogRouter.delete('/:id', async (request, response, next) => {
+    if (blogToDelete.user.toString() !== decodedToken.id.toString()) {
+      response.status(401).json({ error: 'You do not have permission to delete this blog' })
+    }
+
     await Blog.findByIdAndRemove(request.params.id)
     response.status(204).end()
 })
+
+
+blogRouter.put('/:id', async (request, response, next) => {
+  const body = request.body
+
+  const blog = {
+    title: body.title,
+    author: body.author,
+    url: body.url,
+    likes: body.likes,
+    id: body.id
+  }
+
+  await Blog.findByIdAndUpdate(request.params.id, blog)
+  response.status(200).end()
+})
+
+/* voi poistaakkin, vanha tapa joka jo päivitetty
+const getTokenFrom = request => {
+  const authorization = request.get('authorization')
+  if (authorization && authorization.toLowerCase().startsWith('bearer ')) {
+    return authorization.substring(7)
+  }
+  return null
+}
+*/
 
 module.exports = blogRouter
